Change entities
Admin, Organizer, Distributor extend User

+++ Project_tickets/src/main/java/bg/tuvarna/sit/group5/tickets/data/repositories/DAORepository.java
@@ -7,7 +7,7 @@ public interface DAORepository<T> {
     void save(T obj);
     void update(T obj);
     void delete(T obj);
-    T getById(Long id);
+    T getById(int id);
     List<T> getAll();
 
 }
