+++ Project_tickets/src/main/java/bg/tuvarna/sit/group5/tickets/data/repositories/DistributorEventRepository.java
@@ -76,7 +76,7 @@ public class DistributorEventRepository implements DAORepository<DistribEvent>{
     }
 
     @Override
-    public DistribEvent getById(Long id) {
+    public DistribEvent getById(int id) {
         Session session=Connection.openSession();
         Transaction transaction=session.beginTransaction();
         DistribEvent retDistEv;
